audio: QTI resampler supports for 32 bit input format
Configure QTI resampler as 32 bit input format to avoid
losing precision.

audioflinger: add channel count check to use QTI resampler

CRs-Fixed: 940287

+++ services/audioflinger/AudioResamplerQTI.h
@@ -35,7 +35,7 @@ public:
     size_t getNumInSample(size_t outFrameCount);
 
     int16_t *mState;
-    int16_t *mTmpBuf;
+    int32_t *mTmpBuf;
     int32_t *mResamplerOutBuf;
     size_t mFrameIndex;
     size_t stateSize;
